packages/core/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.3.1
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
alsa-lib=1.2.14=hb9d3cd8_0
asttokens=3.0.0=pyhd8ed1ab_1
brotli=1.1.0=hb9d3cd8_2
brotli-bin=1.1.0=hb9d3cd8_2
bzip2=1.0.8=h4bc722e_7
ca-certificates=2025.4.26=hbd8a1cb_0
cairo=1.18.4=h3394656_0
colorama=0.4.6=pyhd8ed1ab_1
comm=0.2.2=pyhd8ed1ab_1
contourpy=1.3.2=py311hd18a35c_0
cycler=0.12.1=pyhd8ed1ab_1
cyrus-sasl=2.1.27=h54b06d7_7
dbus=1.13.6=h5008d03_3
debugpy=1.8.14=py311hfdbb021_0
decorator=5.2.1=pyhd8ed1ab_0
double-conversion=3.3.1=h5888daf_0
exceptiongroup=1.3.0=pyhd8ed1ab_0
executing=2.2.0=pyhd8ed1ab_0
expat=2.7.0=h5888daf_0
filelock=3.13.1=pypi_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_3
fontconfig=2.15.0=h7e30c49_1
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.58.0=py311h2dc5d0c_0
freetype=2.13.3=ha770c72_1
fsspec=2024.6.1=pypi_0
graphite2=1.3.13=h59595ed_1003
harfbuzz=11.2.1=h3beb420_0
icu=75.1=he02047a_0
importlib-metadata=8.6.1=pyha770c72_0
ipykernel=6.29.5=pyh3099207_0
ipython=9.2.0=pyhfb0248b_0
ipython_pygments_lexers=1.1.1=pyhd8ed1ab_0
ipywidgets=8.1.7=pyhd8ed1ab_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.4=pypi_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.7.2=pyh31011fe_1
jupyterlab_widgets=3.0.15=pyhd8ed1ab_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.7=py311hd18a35c_0
krb5=1.21.3=h659f571_0
lcms2=2.17=h717163a_0
ld_impl_linux-64=2.43=h712a8e2_4
lerc=4.0.0=h0aef613_1
libblas=3.9.0=31_h59b9bed_openblas
libbrotlicommon=1.1.0=hb9d3cd8_2
libbrotlidec=1.1.0=hb9d3cd8_2
libbrotlienc=1.1.0=hb9d3cd8_2
libcblas=3.9.0=31_he106b2a_openblas
libclang-cpp20.1=20.1.5=default_h1df26ce_0
libclang13=20.1.5=default_he06ed0a_0
libcups=2.3.3=h4637d8d_4
libdeflate=1.24=h86f0d12_0
libdrm=2.4.124=hb9d3cd8_0
libedit=3.1.20250104=pl5321h7949ede_0
libegl=1.7.0=ha4b6fd6_2
libexpat=2.7.0=h5888daf_0
libffi=3.4.6=h2dba641_1
libfreetype=2.13.3=ha770c72_1
libfreetype6=2.13.3=h48d6fc4_1
libgcc=15.1.0=h767d61c_2
libgcc-ng=15.1.0=h69a702a_2
libgfortran=15.1.0=h69a702a_2
libgfortran5=15.1.0=hcea5267_2
libgl=1.7.0=ha4b6fd6_2
libglib=2.84.1=h3618099_1
libglvnd=1.7.0=ha4b6fd6_2
libglx=1.7.0=ha4b6fd6_2
libgomp=15.1.0=h767d61c_2
libiconv=1.18=h4ce23a2_1
libjpeg-turbo=3.1.0=hb9d3cd8_0
liblapack=3.9.0=31_h7ac8fdf_openblas
libllvm20=20.1.5=he9d0ab4_0
liblzma=5.8.1=hb9d3cd8_1
libnsl=2.0.1=hd590300_0
libntlm=1.8=hb9d3cd8_0
libopenblas=0.3.29=pthreads_h94d23a6_0
libopengl=1.7.0=ha4b6fd6_2
libpciaccess=0.18=hd590300_0
libpng=1.6.47=h943b412_0
libpq=17.5=h27ae623_0
libsodium=1.0.20=h4ab18f5_0
libsqlite=3.49.2=hee588c1_0
libstdcxx=15.1.0=h8f9b012_2
libstdcxx-ng=15.1.0=h4852527_2
libtiff=4.7.0=hf01ce69_5
libuuid=2.38.1=h0b41bf4_0
libwebp-base=1.5.0=h851e524_0
libxcb=1.17.0=h8a09558_0
libxcrypt=4.4.36=hd590300_1
libxkbcommon=1.9.2=h65c71a3_0
libxml2=2.13.8=h4bc477f_0
libxslt=1.1.39=h76b75d6_0
libzlib=1.3.1=hb9d3cd8_2
markupsafe=2.1.5=pypi_0
matplotlib=3.10.3=py311h38be061_0
matplotlib-base=3.10.3=py311h2b939e6_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mpmath=1.3.0=pypi_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.5=h2d0b736_3
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.3=pypi_0
numpy=2.1.2=pypi_0
nvidia-cublas-cu12=12.4.5.8=pypi_0
nvidia-cuda-cupti-cu12=12.4.127=pypi_0
nvidia-cuda-nvrtc-cu12=12.4.127=pypi_0
nvidia-cuda-runtime-cu12=12.4.127=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.2.1.3=pypi_0
nvidia-cufile-cu12=1.13.0.11=pypi_0
nvidia-curand-cu12=10.3.5.147=pypi_0
nvidia-cusolver-cu12=11.6.1.9=pypi_0
nvidia-cusparse-cu12=12.3.1.170=pypi_0
nvidia-cusparselt-cu12=0.6.2=pypi_0
nvidia-nccl-cu12=2.21.5=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.4.127=pypi_0
openjpeg=2.5.3=h5fbd93e_0
openldap=2.6.9=he970967_0
openssl=3.5.0=h7b32b05_1
packaging=25.0=pyh29332c3_1
pandas=2.2.3=py311h7db5c69_3
parso=0.8.4=pyhd8ed1ab_1
pcre2=10.45=hc749103_0
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.0.0=pypi_0
pip=25.1.1=pyh8b19718_0
pixman=0.46.0=h29eaf8c_0
platformdirs=4.3.8=pyhe01879c_0
prompt-toolkit=3.0.51=pyha770c72_0
psutil=7.0.0=py311h9ecbd09_0
pthread-stubs=0.4=hb9d3cd8_1002
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
pygments=2.19.1=pyhd8ed1ab_0
pyparsing=3.2.3=pyhd8ed1ab_1
pyside6=6.9.0=py311h9053184_0
python=3.11.12=h9e4cc4f_0_cpython
python-dateutil=2.9.0.post0=pyhff2d567_1
python-tzdata=2025.2=pyhd8ed1ab_0
python_abi=3.11=7_cp311
pytz=2025.2=pyhd8ed1ab_0
pyzmq=26.4.0=py311h7deb3e3_0
qhull=2020.2=h434a139_5
qt6-main=6.9.0=h0384650_3
readline=8.2=h8c095d6_2
scipy=1.15.2=py311h8f841c2_0
setuptools=80.1.0=pyhff2d567_0
six=1.17.0=pyhd8ed1ab_0
stack_data=0.6.3=pyhd8ed1ab_1
sympy=1.13.1=pypi_0
tk=8.6.13=noxft_h4845f30_101
torch=2.6.0+cu124=pypi_0
torchaudio=2.6.0+cu124=pypi_0
torchvision=0.21.0+cu124=pypi_0
tornado=6.5=py311h9ecbd09_0
tqdm=4.67.1=pyhd8ed1ab_1
traitlets=5.14.3=pyhd8ed1ab_1
triton=3.2.0=pypi_0
typing-extensions=4.12.2=pypi_0
typing_extensions=4.13.2=pyh29332c3_0
tzdata=2025b=h78e105d_0
unicodedata2=16.0.0=py311h9ecbd09_0
wayland=1.23.1=h3e06ad9_1
wcwidth=0.2.13=pyhd8ed1ab_1
wheel=0.45.1=pyhd8ed1ab_1
widgetsnbextension=4.0.14=pyhd8ed1ab_0
xcb-util=0.4.1=hb711507_2
xcb-util-cursor=0.1.5=hb9d3cd8_0
xcb-util-image=0.4.0=hb711507_2
xcb-util-keysyms=0.4.1=hb711507_0
xcb-util-renderutil=0.3.10=hb711507_0
xcb-util-wm=0.4.2=hb711507_0
xkeyboard-config=2.44=hb9d3cd8_0
xorg-libice=1.1.2=hb9d3cd8_0
xorg-libsm=1.2.6=he73a12e_0
xorg-libx11=1.8.12=h4f16b4b_0
xorg-libxau=1.0.12=hb9d3cd8_0
xorg-libxcomposite=0.4.6=hb9d3cd8_2
xorg-libxcursor=1.2.3=hb9d3cd8_0
xorg-libxdamage=1.1.6=hb9d3cd8_0
xorg-libxdmcp=1.1.5=hb9d3cd8_0
xorg-libxext=1.3.6=hb9d3cd8_0
xorg-libxfixes=6.0.1=hb9d3cd8_0
xorg-libxi=1.8.2=hb9d3cd8_0
xorg-libxrandr=1.5.4=hb9d3cd8_0
xorg-libxrender=0.9.12=hb9d3cd8_0
xorg-libxtst=1.2.5=hb9d3cd8_3
xorg-libxxf86vm=1.1.6=hb9d3cd8_0
zeromq=4.3.5=h3b0a872_7
zipp=3.21.0=pyhd8ed1ab_1
zstd=1.5.7=hb8e6e7a_2
